Update version and improve HttpHeaders handling
The application version in build.gradle.kts has been incremented from 1.0.5 to 1.0.6. Additionally, HttpHeaders handling has been modified to exclude technical headers when constructing a headers string for logging purposes. The redundant getHeaders method in RestTemplateLoggingInterceptor.kt was removed and replaced with a call to the new getHeadersString method in RestTemplateConfigHeaders.kt.